Ajoute un datafile à un dataset

Ajoute un datafile à un dataset

Usage

add_datafile(
  dataset,
  title,
  description,
  file_name,
  millesime = NULL,
  published = NULL,
  temporal_coverage_start = NULL,
  temporal_coverage_end = NULL,
  legal_notice = "SDES",
  date_diffusion = NULL,
  quiet = NULL
)

Arguments

dataset

l'id d'un dataset, un objet dido_dataset(), dido_datafile() ou dido_job()

title

le titre du datafile

description

la description du datafile

file_name

le nom du fichier à charger

millesime

le millesime (YYYY-MM). Par défaut prendre la valeur YYYY-MM avec l'année courante et le mois courant

published

la date de publication du fichier, si non précisée, prend la date du jour.

temporal_coverage_start

optionnel, la date de début de couverture du fichier de données au format AAAA-MM-JJ

temporal_coverage_end

optionnel, la date de fin de couverture du fichier de données au format AAAA-MM-JJ

legal_notice

mention légale, par défaut "SDES"

date_diffusion

la date/heure à laquelle le fichier sera accessible au format ISO 8601 (2021-10-01T08:00:00Z). Si non précisée prend la date du jour à minuit, le fichier est donc immédiatement accessible.

quiet

quand TRUE ou que l'option dido_quiet est à TRUE supprime les messages d'information, NULL par défaut

Valeur de retour

un objet dido_job()

Voir également

Exemples

if (FALSE) {
datafile <- add_datafile(
  dataset = dataset$id,
  title = "titre",
  description = "description",
  token_file = file_id
)
}