Ajoute ou modifie un datafile

met à jour le datafile avec le même titre s'il existe sinon ajoute un datafile

Usage

add_or_update_datafile(
  dataset,
  title,
  description,
  file_name,
  millesime = format(Sys.time(), "%Y-%m"),
  published = format(Sys.time(), "%Y-%m-%d"),
  temporal_coverage_start = NULL,
  temporal_coverage_end = NULL,
  legal_notice = "SDES",
  date_diffusion = format(Sys.time(), "%Y-%m-%d %H:%M:%S"),
  keep_old_millesimes = Inf,
  on_existing_millesime = "fail",
  check_file_date = FALSE,
  quiet = NULL
)

Arguments

dataset

l'id d'un dataset, un objet dido_dataset(), dido_datafile() ou dido_job()

title

le titre du datafile

description

la description du datafile

file_name

le nom du fichier à charger

millesime

le millesime (AAAA-MM). Par défaut AAAA-MM avec l'année courante et le mois courant

published

la date de publication du fichier, si non précisée, prend la date du jour.

temporal_coverage_start

optionnel, la date de début de couverture du fichier de données au format AAAA-MM-JJ

temporal_coverage_end

optionnel, la date de fin de couverture du fichier de données au format AAAA-MM-JJ

legal_notice

les mentions légales, par défaut "SDES"

date_diffusion

(optionnel) la date/heure à laquelle le fichier devra être rendu accessible à la diffusion au format "AAAA-MM-JJ HH:MM:SS". Si la time zone n'est pas précisée, l'heure locale est utilisée.

Si cette date est dans le passé, les données sont immédiatement accessibles, si elle est dans le futur, les données ne seront accessibles qu'à cette date/heure.

Si non précisée prend la date/heure courante (heure locale), les données sont donc immédiatement accessibles.

keep_old_millesimes

nombre d'ancien millésimes à conserver, les autres sont supprimés.

on_existing_millesime

skip/fail : action à faire quand le millésime existe déjà. Peut-être :

"skip" : retourne immédiatement après avoir affiché un message

"fail" : lève une exception de class existing_millesime

check_file_date

TRUE/FALSE, Si TRUE met à jour le datafile uniquement si le fichier est plus récent que le last_modified du datafile

Valeur de retour

un objet dido_job() ou NULL si aucune création/mise à jour n'a eu lieu

Voir également

Exemples

library(dplyr, warn.conflicts = FALSE)

dataset <- add_or_update_dataset(
  title = "Un dataset pour les add_or_update_datafiles",
  description = "Description des données statistiques",
  topic = "Transports",
  frequency = "unknown",
)

add_or_update_datafile(
  dataset = dataset,
  title = "titre",
  description = "description",
  file_name = dido_example("augmente.csv"),
  millesime = "2021-10",
)
#> Error in (function (cnd) {    {        .__handler_frame__. <- TRUE        .__setup_frame__. <- frame    }    out <- handlers[[1L]](cnd)    if (!inherits(out, "rlang_zap"))         throw(out)})(structure(list(message = "\033[31m✖\033[39m Erreur: Le millésime que vous souhaitez ajouter existe déjà: 400\n\033[34mℹ\033[39m url: http://api.alimentation.dido.fr/v1/datasets/635026e6412fe41bf7074f86/datafiles/d532134d-d674-4a0b-ac15-988be555f514",     trace = structure(list(call = list(deploy_to_branch(), build_site_github_pages(pkg,         ..., clean = clean), build_site(pkg, preview = FALSE,         install = install, new_process = new_process, ...), build_site_local(pkg = pkg,         examples = examples, run_dont_run = run_dont_run, seed = seed,         lazy = lazy, override = override, preview = preview,         devel = devel), build_reference(pkg, lazy = lazy, examples = examples,         run_dont_run = run_dont_run, seed = seed, override = override,         preview = FALSE, devel = devel), purrr::map(topics, build_reference_topic,         pkg = pkg, lazy = lazy, examples_env = examples_env,         run_dont_run = run_dont_run), .f(.x[[i]], ...), withCallingHandlers(data_reference_topic(topic,         pkg, examples_env = examples_env, run_dont_run = run_dont_run),         error = function(err) {            msg <- c(paste0("Failed to parse Rd in ", topic$file_in),                 i = err$message)            abort(msg, parent = err)        }), data_reference_topic(topic, pkg, examples_env = examples_env,         run_dont_run = run_dont_run), run_examples(tags$tag_examples[[1]],         env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),         topic = tools::file_path_sans_ext(topic$file_in), run_dont_run = run_dont_run),         highlight_examples(code, topic, env = env), downlit::evaluate_and_highlight(code,             fig_save = fig_save_topic, env = child_env(env),             output_handler = evaluate::new_output_handler(value = pkgdown_print)),         evaluate::evaluate(code, child_env(env), new_device = TRUE,             output_handler = output_handler), evaluate_call(expr,             parsed$src[[i]], envir = envir, enclos = enclos,             debug = debug, last = i == length(out), use_try = stop_on_error !=                 2L, keep_warning = keep_warning, keep_message = keep_message,             output_handler = output_handler, include_timing = include_timing),         timing_fn(handle(ev <- withCallingHandlers(withVisible(eval_with_user_handlers(expr,             envir, enclos, user_handlers)), warning = wHandler,             error = eHandler, message = mHandler))), handle(ev <- withCallingHandlers(withVisible(eval_with_user_handlers(expr,             envir, enclos, user_handlers)), warning = wHandler,             error = eHandler, message = mHandler)), try(f, silent = TRUE),         tryCatch(expr, error = function(e) {            call <- conditionCall(e)            if (!is.null(call)) {                if (identical(call[[1L]], quote(doTryCatch)))                   call <- sys.call(-4L)                dcall <- deparse(call, nlines = 1L)                prefix <- paste("Error in", dcall, ": ")                LONG <- 75L                sm <- strsplit(conditionMessage(e), "\n")[[1L]]                w <- 14L + nchar(dcall, type = "w") + nchar(sm[1L],                   type = "w")                if (is.na(w))                   w <- 14L + nchar(dcall, type = "b") + nchar(sm[1L],                     type = "b")                if (w > LONG)                   prefix <- paste0(prefix, "\n  ")            }            else prefix <- "Error : "            msg <- paste0(prefix, conditionMessage(e), "\n")            .Internal(seterrmessage(msg[1L]))            if (!silent && isTRUE(getOption("show.error.messages"))) {                cat(msg, file = outFile)                .Internal(printDeferredWarnings())            }            invisible(structure(msg, class = "try-error", condition = e))        }), tryCatchList(expr, classes, parentenv, handlers),         tryCatchOne(expr, names, parentenv, handlers[[1L]]),         doTryCatch(return(expr), name, parentenv, handler), withCallingHandlers(withVisible(eval_with_user_handlers(expr,             envir, enclos, user_handlers)), warning = wHandler,             error = eHandler, message = mHandler), withVisible(eval_with_user_handlers(expr,             envir, enclos, user_handlers)), eval_with_user_handlers(expr,             envir, enclos, user_handlers), eval(expr, envir,             enclos), eval(expr, envir, enclos), add_or_update_datafile(dataset = dataset,             title = "titre", description = "description", file_name = dido_example("augmente.csv"),             millesime = "2021-10", ), try_fetch({            job_result <- add_millesime(datafile = datafile,                 file_name = file_name, millesime = millesime,                 date_diffusion = date_diffusion)        }, millesime_exists = function(cnd) {            msg <- (c("Le millesime existe déjà :", i = glue::glue("dataset: {dataset$title}"),                 i = glue::glue("datafile: {datafile$title}"),                 i = glue::glue("millesime: {millesime}")))            if (on_existing_millesime == "skip") {                with_didoscalim_verbosity("info", {                  didoscalim_info(msg)                })                return(NULL)            }            else if (on_existing_millesime == "fail") {                didoscalim_abort(cnd$message, class = class(cnd),                   call = caller_env())            }        }), withCallingHandlers(expr, millesime_exists = function(cnd) {            {                .__handler_frame__. <- TRUE                .__setup_frame__. <- frame            }            out <- handlers[[1L]](cnd)            if (!inherits(out, "rlang_zap"))                 throw(out)        }), add_millesime(datafile = datafile, file_name = file_name,             millesime = millesime, date_diffusion = date_diffusion),         tryCatch({            job <- dido_api(method = "POST", path = url, body = body)            job_result <- dido_job(wait_for_job(job$id))        }, error = function(cnd) {            if (grepl("millésime.*existe déjà", cnd$message)) {                class <- c("millesime_exists", class(cnd))                abort(cnd$message, class = class, call = caller_env())            }            else {                didoscalim_abort(parent = cnd)            }        }), tryCatchList(expr, classes, parentenv, handlers),         tryCatchOne(expr, names, parentenv, handlers[[1L]]),         value[[3L]](cond), abort(cnd$message, class = class,             call = caller_env())), parent = c(0L, 1L, 2L, 3L,     4L, 5L, 6L, 7L, 7L, 9L, 10L, 11L, 12L, 13L, 14L, 14L, 16L,     17L, 18L, 19L, 20L, 14L, 14L, 14L, 24L, 25L, 26L, 27L, 28L,     27L, 30L, 31L, 32L, 33L, 34L), visible = c(TRUE, TRUE, TRUE,     TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,     TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,     TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,     FALSE, FALSE), namespace = c("pkgdown", "pkgdown", "pkgdown",     "pkgdown", "pkgdown", "purrr", "pkgdown", "base", "pkgdown",     "pkgdown", "pkgdown", "downlit", "evaluate", "evaluate",     "evaluate", "evaluate", "base", "base", "base", "base", "base",     "base", "base", "evaluate", "base", "base", "didoscalim",     "rlang", "base", "didoscalim", "base", "base", "base", NA,     "rlang"), scope = c("::", "::", "::", ":::", "::", "::",     "local", "::", ":::", ":::", ":::", "::", "::", ":::", "local",     "local", "::", "::", "local", "local", "local", "::", "::",     ":::", "::", "::", "::", "::", "::", "::", "::", "local",     "local", NA, "::"), error_frame = c(FALSE, FALSE, FALSE,     FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,     FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,     FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,     FALSE, FALSE, TRUE, FALSE, FALSE)), row.names = c(NA, -35L    ), version = 2L, class = c("rlang_trace", "rlib_trace", "tbl",     "data.frame")), parent = NULL, call = tryCatchOne(expr, names,         parentenv, handlers[[1L]])), class = c("millesime_exists", "api_error", "http_error_400", "didoscalim_error", "rlang_error", "error", "condition", "rlang_error", "error", "condition"))): ✖ Erreur: Le millésime que vous souhaitez ajouter existe déjà: 400
#> ℹ url: http://api.alimentation.dido.fr/v1/datasets/635026e6412fe41bf7074f86/datafiles/d532134d-d674-4a0b-ac15-988be555f514

add_or_update_datafile(
  dataset = dataset,
  title = "titre",
  description = "description",
  file_name = dido_example("augmente.csv"),
  temporal_coverage_start = "2021-01-01",
  temporal_coverage_end = "2021-12-31",
  millesime = "2022-10",
)
#> Error in (function (cnd) {    {        .__handler_frame__. <- TRUE        .__setup_frame__. <- frame    }    out <- handlers[[1L]](cnd)    if (!inherits(out, "rlang_zap"))         throw(out)})(structure(list(message = "\033[31m✖\033[39m Erreur: Le millésime que vous souhaitez ajouter existe déjà: 400\n\033[34mℹ\033[39m url: http://api.alimentation.dido.fr/v1/datasets/635026e6412fe41bf7074f86/datafiles/d532134d-d674-4a0b-ac15-988be555f514",     trace = structure(list(call = list(deploy_to_branch(), build_site_github_pages(pkg,         ..., clean = clean), build_site(pkg, preview = FALSE,         install = install, new_process = new_process, ...), build_site_local(pkg = pkg,         examples = examples, run_dont_run = run_dont_run, seed = seed,         lazy = lazy, override = override, preview = preview,         devel = devel), build_reference(pkg, lazy = lazy, examples = examples,         run_dont_run = run_dont_run, seed = seed, override = override,         preview = FALSE, devel = devel), purrr::map(topics, build_reference_topic,         pkg = pkg, lazy = lazy, examples_env = examples_env,         run_dont_run = run_dont_run), .f(.x[[i]], ...), withCallingHandlers(data_reference_topic(topic,         pkg, examples_env = examples_env, run_dont_run = run_dont_run),         error = function(err) {            msg <- c(paste0("Failed to parse Rd in ", topic$file_in),                 i = err$message)            abort(msg, parent = err)        }), data_reference_topic(topic, pkg, examples_env = examples_env,         run_dont_run = run_dont_run), run_examples(tags$tag_examples[[1]],         env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),         topic = tools::file_path_sans_ext(topic$file_in), run_dont_run = run_dont_run),         highlight_examples(code, topic, env = env), downlit::evaluate_and_highlight(code,             fig_save = fig_save_topic, env = child_env(env),             output_handler = evaluate::new_output_handler(value = pkgdown_print)),         evaluate::evaluate(code, child_env(env), new_device = TRUE,             output_handler = output_handler), evaluate_call(expr,             parsed$src[[i]], envir = envir, enclos = enclos,             debug = debug, last = i == length(out), use_try = stop_on_error !=                 2L, keep_warning = keep_warning, keep_message = keep_message,             output_handler = output_handler, include_timing = include_timing),         timing_fn(handle(ev <- withCallingHandlers(withVisible(eval_with_user_handlers(expr,             envir, enclos, user_handlers)), warning = wHandler,             error = eHandler, message = mHandler))), handle(ev <- withCallingHandlers(withVisible(eval_with_user_handlers(expr,             envir, enclos, user_handlers)), warning = wHandler,             error = eHandler, message = mHandler)), try(f, silent = TRUE),         tryCatch(expr, error = function(e) {            call <- conditionCall(e)            if (!is.null(call)) {                if (identical(call[[1L]], quote(doTryCatch)))                   call <- sys.call(-4L)                dcall <- deparse(call, nlines = 1L)                prefix <- paste("Error in", dcall, ": ")                LONG <- 75L                sm <- strsplit(conditionMessage(e), "\n")[[1L]]                w <- 14L + nchar(dcall, type = "w") + nchar(sm[1L],                   type = "w")                if (is.na(w))                   w <- 14L + nchar(dcall, type = "b") + nchar(sm[1L],                     type = "b")                if (w > LONG)                   prefix <- paste0(prefix, "\n  ")            }            else prefix <- "Error : "            msg <- paste0(prefix, conditionMessage(e), "\n")            .Internal(seterrmessage(msg[1L]))            if (!silent && isTRUE(getOption("show.error.messages"))) {                cat(msg, file = outFile)                .Internal(printDeferredWarnings())            }            invisible(structure(msg, class = "try-error", condition = e))        }), tryCatchList(expr, classes, parentenv, handlers),         tryCatchOne(expr, names, parentenv, handlers[[1L]]),         doTryCatch(return(expr), name, parentenv, handler), withCallingHandlers(withVisible(eval_with_user_handlers(expr,             envir, enclos, user_handlers)), warning = wHandler,             error = eHandler, message = mHandler), withVisible(eval_with_user_handlers(expr,             envir, enclos, user_handlers)), eval_with_user_handlers(expr,             envir, enclos, user_handlers), eval(expr, envir,             enclos), eval(expr, envir, enclos), add_or_update_datafile(dataset = dataset,             title = "titre", description = "description", file_name = dido_example("augmente.csv"),             temporal_coverage_start = "2021-01-01", temporal_coverage_end = "2021-12-31",             millesime = "2022-10", ), try_fetch({            job_result <- add_millesime(datafile = datafile,                 file_name = file_name, millesime = millesime,                 date_diffusion = date_diffusion)        }, millesime_exists = function(cnd) {            msg <- (c("Le millesime existe déjà :", i = glue::glue("dataset: {dataset$title}"),                 i = glue::glue("datafile: {datafile$title}"),                 i = glue::glue("millesime: {millesime}")))            if (on_existing_millesime == "skip") {                with_didoscalim_verbosity("info", {                  didoscalim_info(msg)                })                return(NULL)            }            else if (on_existing_millesime == "fail") {                didoscalim_abort(cnd$message, class = class(cnd),                   call = caller_env())            }        }), withCallingHandlers(expr, millesime_exists = function(cnd) {            {                .__handler_frame__. <- TRUE                .__setup_frame__. <- frame            }            out <- handlers[[1L]](cnd)            if (!inherits(out, "rlang_zap"))                 throw(out)        }), add_millesime(datafile = datafile, file_name = file_name,             millesime = millesime, date_diffusion = date_diffusion),         tryCatch({            job <- dido_api(method = "POST", path = url, body = body)            job_result <- dido_job(wait_for_job(job$id))        }, error = function(cnd) {            if (grepl("millésime.*existe déjà", cnd$message)) {                class <- c("millesime_exists", class(cnd))                abort(cnd$message, class = class, call = caller_env())            }            else {                didoscalim_abort(parent = cnd)            }        }), tryCatchList(expr, classes, parentenv, handlers),         tryCatchOne(expr, names, parentenv, handlers[[1L]]),         value[[3L]](cond), abort(cnd$message, class = class,             call = caller_env())), parent = c(0L, 1L, 2L, 3L,     4L, 5L, 6L, 7L, 7L, 9L, 10L, 11L, 12L, 13L, 14L, 14L, 16L,     17L, 18L, 19L, 20L, 14L, 14L, 14L, 24L, 25L, 26L, 27L, 28L,     27L, 30L, 31L, 32L, 33L, 34L), visible = c(TRUE, TRUE, TRUE,     TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,     TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,     TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,     FALSE, FALSE), namespace = c("pkgdown", "pkgdown", "pkgdown",     "pkgdown", "pkgdown", "purrr", "pkgdown", "base", "pkgdown",     "pkgdown", "pkgdown", "downlit", "evaluate", "evaluate",     "evaluate", "evaluate", "base", "base", "base", "base", "base",     "base", "base", "evaluate", "base", "base", "didoscalim",     "rlang", "base", "didoscalim", "base", "base", "base", NA,     "rlang"), scope = c("::", "::", "::", ":::", "::", "::",     "local", "::", ":::", ":::", ":::", "::", "::", ":::", "local",     "local", "::", "::", "local", "local", "local", "::", "::",     ":::", "::", "::", "::", "::", "::", "::", "::", "local",     "local", NA, "::"), error_frame = c(FALSE, FALSE, FALSE,     FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,     FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,     FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,     FALSE, FALSE, TRUE, FALSE, FALSE)), row.names = c(NA, -35L    ), version = 2L, class = c("rlang_trace", "rlib_trace", "tbl",     "data.frame")), parent = NULL, call = tryCatchOne(expr, names,         parentenv, handlers[[1L]])), class = c("millesime_exists", "api_error", "http_error_400", "didoscalim_error", "rlang_error", "error", "condition", "rlang_error", "error", "condition"))): ✖ Erreur: Le millésime que vous souhaitez ajouter existe déjà: 400
#> ℹ url: http://api.alimentation.dido.fr/v1/datasets/635026e6412fe41bf7074f86/datafiles/d532134d-d674-4a0b-ac15-988be555f514