Créé un dataset

Créé un dataset

create_dataset(
  title,
  description,
  topic,
  frequency,
  tags = NULL,
  frequency_date = NULL,
  granularity = NULL,
  zones = NULL,
  organization = NULL,
  license = "fr-lo",
  temporal_coverage_start = NULL,
  temporal_coverage_end = NULL,
  caution = NULL,
  quiet = FALSE
)

Arguments

title
le titre du dataset

description
la description du dataset

topic
le thème du dataset. Doit être dans "Environnement", "Énergie", "Transports", "Logement", "Changement climatique"

frequency
la fréquence de publication des données

tags
la liste des mots clefs. Si le fichier est déjà publié sur le site SDES, on reprend les mêmes. L'ensemble des mots clefs est disponible sur https://data.statistiques.developpement-durable.gouv.fr/dido/api/v1/referentiels/tags/csv?withColumnName=true&withColumnDescription=false.

frequency_date
la date de prochaine publication.

granularity
la granularité du jeu de données. Les plus utilisés sont "fr:region", "fr:departement", "fr:epci", "fr:commune", "fr:iris" et pour les données à l'adresse "poi"

zones
la zone couverte par le jeu de données. country:fr pour France entière, country-subset:fr:metro pour la métropole et country-subset:fr:drom pour les DROM

organization
l'id de l'organisation sous laquelle vous souhaitez publier le dataset, si vous n'appartenez qu'à une seule organisation, didoscalim la prendra par défaut.

license
la licence des données. "fr-lo" par défaut

temporal_coverage_start
la date de début de couverture du jeux de données

temporal_coverage_end
la date de fin de couverture du jeux de données

caution
Les précautions à prendre avec ce jeu de données. Par exemple : "certaines données peuvent être secrétisées"

quiet
Si TRUE n'affiche pas les messages. Défaut à FALSE

Value

un objet dido_dataset(). Ce dernier sert pour créer par la suite les datafiles et les millésimes.

Examples

dataset <- create_dataset(
  title = "test",
  description = "test",
  topic = "Transports",
  frequency = "unknown"
)
#> Environnement DiDo actif : DEV
#> dataset `test` créé