L'objet dido_datafile

L'objet dido_datafile et retourné par get_datafile() et peut-être utilisé en paramètre par les fonctions de manipulation de dataset, datafile, millésime et attachement.

Il a les champs suivants :

rid l'identifiant du datafile

title

description

millesimes le nb de millésimes dans le datafile

created_at

last_modified

published

url

temporal_coverage un liste avec deux champs start et end

legal_notice la licence s'appliquant au fichier de données

millesimes_info comprenant une liste des millésimes rattachés

le dernier champ est supprimé par clean_metadata()

L'objet a également un attribut id avec l'identifiant du dataset auquel il appartient

Cette fonction est utilisée par add_datafile()

dido_datafile(
  dataset,
  title,
  description,
  millesime = NULL,
  published = NULL,
  temporal_coverage_start = NULL,
  temporal_coverage_end = NULL,
  legal_notice = "SDES",
  date_diffusion = NULL
)

Arguments

dataset
l'id d'un dataset, un objet dido_dataset(), dido_datafile() ou dido_job()

title
le titre du datafile

description
la description du datafile

millesime
le millesime (YYYY-MM). Par défaut prendre la valeur YYYY-MM avec l'année courante et le mois courant

published
la date de publication du fichier, si non précisée, prend la date du jour.

temporal_coverage_start
la date de début des données au format YYYY-MM-DD

temporal_coverage_end
la date de fin des données au format YYYY-MM-DD

legal_notice
mention légale, par défaut "SDES"

date_diffusion
la date/heure à laquelle le fichier sera accessible au format ISO 8601 (2021-10-01T08:00:00Z). Si non précisée prend la date du jour à minuit, le fichier est donc immédiatement accessible.

Value

un objet [dido_datafile()]

Examples

if (FALSE) {
datafile <- dido_datafile(
  dataset = "1",
  title = "titre",
  description = "description",
  date_diffusion = "2022-01-01T08:00:00Z"
)
}