Génère les lignes d'entête du CSV augmenté

Génère un dataframe avec les lignes d'entêtes du CSV augmenté comme premières lignes.

dido_csv(
  data,
  params = list(),
  locale = readr::default_locale(),
  cog_year = format(Sys.time(), "%Y")
)

Arguments

data
le dataframe à augmenter

params
une liste nommée décrivant les caractéristiques des colonnes

locale
la locale à utiliser

cog_year
le millésime du COG utilisé si besoin. Par défaut prend l'année en cours

Value

un dataframe avec les 4 lignes de description du csv augmenté

Details

Certains noms de variable sont connus par didoscalim qui génère automatiquement le type et la description. La liste complète de ces variables et des types/descriptions associés est :

nom de la variable	type	description

REGION	cog_region_AAAA	Code de la région

DEPARTEMENT	cog_departement_AAAA	Code du département

COMMUNE	cog_commune_AAAA	Code de la commune

EPCI	cog_epci_AAAA	Code de l'EPCI

IRIS	cog_iris_AAAA	Code de l'IRIS

ANNEE	n/a	Millésimes des données

MOIS	n/a	mois des données

L'année AAAA est par défaut l'année courante, vous pouvez la modifier en passant le paramètre cog_year

See also

Examples

if (FALSE) {
params <- list(
  OPERATEUR = list(description = "L'opérateur", type = "texte"),
  CONSO = list(description = "La consommation", unit = "Mwh")
)
dido_csv(my_tibble, params = params)
}